refactor(diagram): changes according CR
