+++ libs/shared/src/components/diagram/bar-diagram/index.tsx
@@ -1,12 +1,13 @@
-import { DiagramType } from "..";
-import { DIAGRAM_CONFIG } from "./config";
+import { DIAGRAM_CONFIG } from "../../../configs";
+import { DiagramType } from "../../../types";
+import { getDiagramOptions } from "../../../utils";
 import * as Styled from "./styled";
-import { useDiagramOptions } from "./use-diagram-options";
 
 interface BarProps {
   diagramData: { labels: string[]; datasets: any[] };
   title: string;
 }
+
 export const BarDiagram: React.FC<BarProps> = ({ diagramData, title }) => (
-  <Styled.DiagramBar data={diagramData} options={useDiagramOptions(DiagramType.Bar, title)} height={DIAGRAM_CONFIG.height} />
+  <Styled.DiagramBar data={diagramData} options={getDiagramOptions(DiagramType.Bar, title)} height={DIAGRAM_CONFIG.height} />
 );
