+++ libs/shared/src/configs/diagram-config.ts
