+++ libs/shared/src/configs/index.ts
@@ -1,1 +1,2 @@
+export * from "./diagram-config";
 export * from "./param-keys";
