+++ libs/shared/src/types/diagram-type.ts
@@ -1,0 +1,3 @@
+export enum DiagramType {
+  Bar = "Bar"
+}
