+++ libs/shared/src/types/index.ts
@@ -1,5 +1,6 @@
 export * from "./app-routes";
 export * from "./common";
+export * from "./diagram-type";
 export * from "./form-field-type";
 export * from "./http-status";
 export * from "./notification-type";
