+++ libs/shared/src/utils/index.ts
@@ -1,6 +1,7 @@
 export * from "./create-path";
 export * from "./filter-by-keys";
 export * from "./get-app-route";
+export * from "./get-diagram-options";
 export * from "./get-domain-name";
 export * from "./get-theme-config";
 export * from "./lang-fns";
